+++ src/rollem-bot/behaviors/parse-soft.behavior.ts
@@ -5,6 +5,7 @@ import { Config } from "@bot/config";
 import { Injectable } from "injection-js";
 import { Parsers } from "@bot/lib/parsers";
 import { RepliedMessageCache } from "@bot/lib/replied-message-cache";
+import { Storage } from "@storage/storage";
 
 /**
  * Parses things without any prefix, unless a prefix is configured via role-name.
@@ -14,18 +15,19 @@ export class ParseSoftBehavior extends RollBehaviorBase {
   constructor(
     parsers: Parsers,
     config: Config,
+    storage: Storage,
     repliedMessageCache: RepliedMessageCache,
     client: Client,
     logger: Logger,
-  ) { super(parsers, config, repliedMessageCache, client, logger); }
+  ) { super(parsers, config, storage, repliedMessageCache, client, logger); }
 
   protected async register() {
     this.client.on('message', async message => {
-      // avoid doing insane things
+      // avoid doing absurd things
       if (message.author.bot) { return; }
       if (message.author == this.client.user) { return; }
       if (this.repliedMessageCache.hasSeenMessageBefore(message, "soft")) { return; }
-      if (this.shouldDefer(message)) { return; }
+      if (await this.shouldDefer(message)) { return; }
       if (message.content.startsWith('D')) { return; } // apparently D8 is a common emote.
 
       // honor the prefix