+++ src/rollem-language-2/evaluators/make-integer.syntax.spec.ts
@@ -11,7 +11,6 @@ describe('make-integer (syntax)', () => {
     const value = delayedValue(ctx);
     expect(ctx.callCount).to.equal(0);
     expect(value.value).to.equal(5);
-    expect(value.values).to.eql([5]);
     expect(value.pretties).to.equal("5");
     expect(value.dice).to.equal(0);
     expect(value.depth).to.equal(1);
@@ -24,7 +23,6 @@ describe('make-integer (syntax)', () => {
     const value = delayedValue(ctx);
     expect(ctx.callCount).to.equal(0);
     expect(value.value).to.equal(189465);
-    expect(value.values).to.eql([189465]);
     expect(value.pretties).to.equal("189465");
     expect(value.dice).to.equal(0);
     expect(value.depth).to.equal(1);
@@ -34,9 +32,9 @@ describe('make-integer (syntax)', () => {
     const ctx = new TestContext("test");
     const delayedValue = parser.parse("-189465");
     const value = delayedValue(ctx);
+    debugger;
     expect(ctx.callCount).to.equal(0);
     expect(value.value).to.equal(-189465);
-    expect(value.values).to.eql([-189465]);
     expect(value.pretties).to.equal("-189465");
     expect(value.dice).to.equal(0);
     expect(value.depth).to.equal(1);