+++ src/rollem-language-2/evaluators/roll-simple.ts
@@ -1,7 +1,4 @@
-import { Integer } from "@language-v2/types/integer";
-import { OldContainer } from "@language-v2/types";
-import { Delayed } from "@language-v2/types/delayed";
-import { Dice } from "@language-v2/types/dice";
+import { Delayed, Integer, Dice } from "@language-v2/types";
 
 function minFormatter(formatted) {
   return "**" + formatted + "**";
@@ -16,7 +13,7 @@ function dropFormatter(formatted) {
 }
 
 // This is used to configure stylization of the individual die results.
-function dieFormatter(value, size, isKept = true) {
+export function dieFormatter(value, size, isKept = true) {
   let formatted = value
   if (value >= size)
     formatted = maxFormatter(formatted);
@@ -29,12 +26,14 @@ function dieFormatter(value, size, isKept = true) {
   return formatted;
 }
 
-export function rollSimple($$howMany: Delayed<Integer> | null | undefined, $$dieSize: Delayed<Integer>): Delayed<OldContainer> {
+export function rollSimple($$howMany: Delayed<Integer> | null | undefined, $$dieSize: Delayed<Integer>): Delayed<Dice | Integer> {
   return (ctx) => {
     const allRolls: number[] = [];
-    const $howMany = $$howMany ? $$howMany(ctx) : null;
-    const howMany = $howMany?.value ?? 1;
+    const $howMany = $$howMany ? $$howMany(ctx) : Integer.fromNumber(1);
+    const howMany = $howMany.value;
     const $dieSize = $$dieSize(ctx);
+    ctx.trace(`roll-simple: die-size: ${$dieSize}`);
+    ctx.trace(`roll-simple: how-many: ${howMany}`);
 
     for (let i = 0; i < howMany; i++) {
       allRolls.push(ctx.chance.integer({min: 1, max: $dieSize.value}));
@@ -42,13 +41,16 @@ export function rollSimple($$howMany: Delayed<Integer> | null | undefined, $$die
 
     allRolls.sort();
     const sum = allRolls.reduce((accum, cur) => accum + cur, 0);
+
+    // TODO: this formatter should probably preserve the contents of the right hand side if the type is complex
     const pretties = `[${allRolls.map(v => dieFormatter(v, $dieSize.value, true)).join(", ")}] ⟵ ${howMany}d${$dieSize.value}`
     return new Dice({
+      $howMany: $howMany,
+      $dieSize: $dieSize,
       value: sum,
       values: allRolls,
       pretties: pretties,
-      depth: 1,
-      dice: howMany
+      parentValues: [$howMany, $dieSize],
     });
   };
 }