+++ src/storage/entity/User.ts
@@ -1,0 +1,18 @@
+import {Entity, PrimaryGeneratedColumn, Column, CreateDateColumn, UpdateDateColumn, Unique, Index} from "typeorm";
+
+@Entity()
+export class User {
+
+    @PrimaryGeneratedColumn("uuid")
+    id!: string;
+
+    @Column()
+    @Index({ unique: true })
+    discordUserId!: string;
+
+    @CreateDateColumn()
+    created!: Date;
+
+    @UpdateDateColumn()
+    updated!: Date;
+}